Move tests and use toml instead json
