+++ zebra-network/Cargo.toml
@@ -38,6 +38,6 @@ zebra-chain = { path = "../zebra-chain" }
 [dev-dependencies]
 proptest = "0.10"
 proptest-derive = "0.3"
-serde_json = "1"
+toml = "0.5"
 
 zebra-test = { path = "../zebra-test/" }
