+++ zebra-network/src/config/tests.rs
@@ -1,0 +1,22 @@
+use super::Config;
+
+#[test]
+fn parse_config_listen_addr() {
+    let fixtures = vec![
+        ("listen_addr = '0.0.0.0'", "0.0.0.0:8233"),
+        ("listen_addr = '0.0.0.0:9999'", "0.0.0.0:9999"),
+        (
+            "listen_addr = '0.0.0.0'\nnetwork = 'Testnet'",
+            "0.0.0.0:18233",
+        ),
+        (
+            "listen_addr = '0.0.0.0:8233'\nnetwork = 'Testnet'",
+            "0.0.0.0:8233",
+        ),
+    ];
+
+    for (config, value) in fixtures {
+        let config: Config = toml::from_str(config).unwrap();
+        assert_eq!(config.listen_addr.to_string(), value);
+    }
+}
